requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
binance-connector==3.7.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pytz==2024.1
sqlalchemy==2.0.23
alembic==1.13.1


